Create licencBadges list, also change question
