+++ Develop/utils/questions.js
@@ -1,0 +1,110 @@
+// TODO: Create an array of questions for user input
+const questions = [
+    {
+        type: 'input',
+        message: "What is your GitHub username? (No @ needed)",
+        name: 'username',
+        default: 'dallasg84',
+        validate: function (answer) {
+            if (answer.length < 1) {
+                return console.log("A valid GitHub username is required.");
+            }
+            return true;
+        }
+    },
+    {
+        type: 'input',
+        message: "What is the name of your GitHub repo?",
+        name: 'repo',
+        default: 'readme-generator',
+        validate: function (answer) {
+            if (answer.length < 1) {
+                return console.log("A valid GitHub repo is required for a badge.");
+            }
+            return true;
+        }
+    },
+    {
+        type: 'input',
+        message: "What is the title of your project?",
+        name: 'title',
+        default: 'Project Title',
+        validate: function (answer) {
+            if (answer.length < 1) {
+                return console.log("A valid project title is required.");
+            }
+            return true;
+        }
+    },
+    {
+        type: 'input',
+        message: "Write a description of your project.",
+        name: 'description',
+        default: 'Project Description',
+        validate: function (answer) {
+            if (answer.length < 1) {
+                return console.log("A valid project description is required.");
+            }
+            return true;
+        }
+    },
+    {
+        type: "input",
+        name: "userEmail",
+        message: "What is your email?",
+        validate: function (answer) {
+            if (answer.length < 1) {
+                return console.log("A contact email is required.");
+            }
+            return true;
+        }
+    },
+    {
+        type: 'input',
+        message: "If applicable, describe the steps required to install your project for the Installation section.",
+        name: 'installation'
+    },
+    {
+        type: 'input',
+        message: "Provide instructions and examples of your project in use for the Usage section.",
+        name: 'usage'
+    },
+    {
+        type: 'input',
+        message: "If applicable, provide guidelines on how other developers can contribute to your project.",
+        name: 'contributing'
+    },
+    {
+        type: 'input',
+        message: "If applicable, provide any tests written for your application and provide examples on how to run them.",
+        name: 'tests'
+    },
+    {
+        type: "input",
+        name: "URL",
+        message: "What is the URL of the live site?",
+    },
+    {
+        type: "input",
+        name: "repo",
+        message: "What is the URL of the github repo?",
+    },
+    {
+        type: "list",
+        name: "license",
+        message: "Please select the license you used for this project.",
+        choices: [
+          "GNU AGPLv3",
+          "GNU GPLv3",
+          "GNU LGPLv3",
+          "Mozilla",
+          "MIT",
+          "Apache",
+          "Boost",
+        ],
+      },
+];
+
+module.exports = {
+    questions: questions,
+};
